Replace new/unnamed buffer if it was taken over by Niffler
If :Niffler is called in a new buffer created with :new or :tabe then
Vim uses the new buffer for Niffler instead of creating a separate one.
This means that the origin_buffer and the Niffler buffer will be the
same buffer, and attempting to wipeout the Niffler buffer also wipes out
the origin buffer and closes the window that the new buffer was
originally inhabiting.

The fix is to just open a replacement new buffer in this scenario.