Use getbufinfo() to build buffer list. Fixes #47
For terminal buffers Vim displays a different name in the output of
`:ls` than what the buffer's actual name is. This causes Niffler to try
to create a buffer with the incorrect name instead of switching to it.

Use the `getbufinfo()` function, if it's available, instead of parsing
the output of the ls command. This function will ensure that the name of
the buffer we use, display, and attempt to switch to matches up with the
buffer's actual name rather than its display name.