Silence getchar() error in event loop. Fixes #49.
Neovim is throwing E5677 for some reason when calling getchar() in the
keypress_event_loop. This error doesn't seem to actually prevent
getchar() from returning the typed character and silencing it doesn't
seem to have any adverse consequences. This seems like a bug in Neovim
but will have to follow up for more information.